requirements.txt
numpy
matplotlib
python-dotenv
pillow